+++ routes/index.js
@@ -2,12 +2,12 @@ const express = require('express');
 
 const userRouter = require('./userRouter');
 const bookRouter = require('./bookRouter');
-const paymentRouter = require('./paymentRouter.js');
+const orderRouter = require('./orderRouter.js');
 
 const router = express.Router();
 
 router.use('/users', userRouter.router);
 router.use('/books', bookRouter.router);
-router.use('/payments', paymentRouter.router);
+router.use('/orders', orderRouter.router);
 
 module.exports = router;
