+++ routes/orderRouter.js
@@ -1,0 +1,11 @@
+const express = require('express');
+
+const orderController = require('../controllers/orderController.js');
+
+const router = express.Router();
+
+const { validateToken } = require('../middlewares/auth.js');
+
+router.post('/complete', validateToken, orderController.getUserInfo);
+
+module.exports = { router };
