--- routes/paymentRouter.js
@@ -1,11 +1,0 @@
-const express = require('express');
-
-const paymentController = require('../controllers/paymentController.js');
-
-const router = express.Router();
-
-const { validateToken } = require('../middlewares/auth');
-
-router.post('/order', validateToken, paymentController.getUserInfo);
-
-module.exports = { router };
